requirements.txt
six>=1.9.0,<1.10
cmd2>=0.6.8,<0.7
jinja2>=2.7.3,<2.8
pyyaml
